some renaming and bug fixes
