+++ examples/nlp/language_modeling/conf/transformer_lm_config.yaml
@@ -1,26 +1,10 @@
-# Copyright (c) 2020, NVIDIA CORPORATION.  All rights reserved.
-# Copyright 2018 The Google AI Language Team Authors and
-# The HuggingFace Inc. team.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-# Config file for training Transformer language model
+# Config file for training left-to-right Transformer language model
 
 trainer:
-  gpus: 4 # the number of gpus, 0 for CPU
+  gpus: 1 # the number of gpus, 0 for CPU
   num_nodes: 1
-  max_epochs: 300
-  max_steps: 500000 # precedence over max_epochs
+  max_epochs: 2
+  max_steps: 400 # precedence over max_epochs
   accumulate_grad_batches: 1 # accumulates grads every k batches
   amp_level: O2 # O1/O2 for mixed precision
   precision: 16 # Should be set to 16 for O1 and O2, default is 16 as PT ignores it when am_level is O0
@@ -32,9 +16,22 @@ trainer:
   resume_from_checkpoint: null # The path to a checkpoint file to continue the training, restores the whole state including the epoch, step, LR schedulers, apex, etc.
 
 model:
+
+  language_model:
+    tokenizer: CharTokenizer
+    vocab_file: /workspace/datasets/aishell2/vocab.txt
+    hidden_size: 512
+    num_layers: 6
+    num_attn_heads: 8
+    inner_size: 2048
+    max_seq_length: 256
+    embedding_dropout: 0
+    ffn_dropout: 0
+    attn_score_dropout: 0
+    attn_layer_dropout: 0
+
   dataset:
     max_seq_length: 256
-    # parameters shared among all data loaders
     num_workers: 2 # number of workers for data loaders
     drop_last: false # drops the last last batch if it is smaller than the batch size
     pin_memory: false # enables pin_memory feature of the data loaders
@@ -52,30 +49,14 @@ model:
     num_samples: -1 # number of samples to be considered, -1 means all the dataset
     predict_last_k: 64
 
-  language_model:
-    tokenizer: charlevel
-    vocab_file: /workspace/datasets/aishell2/vocab.txt
-    hidden_size: 512
-    num_layers: 6
-    num_attn_heads: 8
-    inner_size: 2048
-    max_seq_length: 256
-    embedding_dropout: 0
-    ffn_dropout: 0
-    attn_score_dropout: 0
-    attn_layer_dropout: 0
-
   optim:
     name: adam
     lr: 1e-4
-    # optimizer arguments
     betas: [0.9, 0.999]
     weight_decay: 0
 
-    # scheduler setup
     sched:
       name: WarmupAnnealing
-      # Scheduler params
       warmup_steps: null
       warmup_ratio: 0.05
       last_epoch: -1
@@ -85,10 +66,10 @@ model:
       reduce_on_plateau: false
 
 exp_manager:
-  root_dir: aishell  # root_dir for your experiment, if None, defaults to "./NeMo_experiments"
-  name: "TransformerLanguageModeling"  # The name of your model
-  create_tensorboard_logger: True  # Whether you want exp_manger to create a tb logger
-  create_checkpoint_callback: True  # Whether you want exp_manager to create a modelcheckpoint callback
+  exp_dir: aishell  # where to store logs and checkpoints
+  name: TransformerLanguageModeling  # The name of your model
+  create_tensorboard_logger: True
+  create_checkpoint_callback: True
 
 hydra:
   run:
