+++ nemo/collections/nlp/data/__init__.py
@@ -13,8 +13,9 @@
 # limitations under the License.
 
 from nemo.collections.nlp.data.data_utils import *
-from nemo.collections.nlp.data.lm_bert_dataset import *
-from nemo.collections.nlp.data.lm_transformer_dataset import *
+from nemo.collections.nlp.data.language_modeling.l2r_lm_dataset import (
+    L2RLanguageModelingDataset,
+)
 from nemo.collections.nlp.data.language_modeling.lm_bert_dataset import (
     BertPretrainingDataset,
     BertPretrainingPreprocessedDataloader,
