+++ nemo/collections/nlp/models/__init__.py
@@ -14,8 +14,8 @@
 
 from nemo.collections.nlp.models.glue_benchmark.glue_benchmark_model import GLUEModel
 from nemo.collections.nlp.models.intent_slot_classification_model import IntentSlotClassificationModel
-from nemo.collections.nlp.models.language_modeling.lm_model import BERTLMModel
-from nemo.collections.nlp.models.language_modeling import TransformerLMModel
+from nemo.collections.nlp.models.language_modeling.bert_lm_model import BERTLMModel
+from nemo.collections.nlp.models.language_modeling.transformer_lm_model import TransformerLMModel
 from nemo.collections.nlp.models.question_answering.qa_model import QAModel
 from nemo.collections.nlp.models.text_classification import TextClassificationModel
 from nemo.collections.nlp.models.token_classification import PunctuationCapitalizationModel, TokenClassificationModel
