+++ nemo/collections/nlp/models/language_modeling/transformer_lm_model.py
@@ -22,13 +22,14 @@ from pytorch_lightning import Trainer
 
 from nemo.collections.common.losses import SmoothedCrossEntropyLoss
 from nemo.collections.common.tokenizers.tokenizer_utils import get_tokenizer
-from nemo.collections.nlp.data import LanguageModelingDataset
+from nemo.collections.nlp.data import L2RLanguageModelingDataset
 from nemo.collections.nlp.modules.common import TokenClassifier
 from nemo.collections.nlp.modules.common.transformer import (
     TransformerEmbedding,
     TransformerEncoder,
     transformer_weights_init,
 )
+from nemo.collections.common.tokenizers import WordTokenizer, CharTokenizer
 from nemo.core.classes.common import typecheck
 from nemo.core.classes.modelPT import ModelPT
 from nemo.utils.decorators import experimental
@@ -44,9 +45,9 @@ class TransformerLMModel(ModelPT):
 
         # shared params for dataset and data loaders
         self.dataset_cfg = cfg.dataset
-
-        self.tokenizer = get_tokenizer(
-            tokenizer_name=cfg.language_model.tokenizer, vocab_file=cfg.language_model.vocab_file,
+        
+        self.tokenizer = globals()[cfg.language_model.tokenizer](
+            vocab_file=cfg.language_model.vocab_file
         )
 
         # make vocabulary size divisible by 8 for fast fp16 training
@@ -153,7 +154,7 @@ class TransformerLMModel(ModelPT):
     def _setup_dataloader_from_config(self, cfg: DictConfig):
         input_file = cfg.file_name
 
-        dataset = LanguageModelingDataset(
+        dataset = L2RLanguageModelingDataset(
             tokenizer=self.tokenizer,
             dataset=cfg.file_name,
             max_seq_length=self.dataset_cfg.max_seq_length,
